requirements.txt
fpdf2==2.7.8
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
ffmpeg-python==0.2.0
openai==1.14.1
streamlit==1.32.2
pydrive==1.3.1